+++ Integrations/VS-Code/CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - `scriptbee.enableLiveUpdates` to enable live updates for project script events of creation, update, and deletion.
+- `scriptbee.enableAutoPush` to enable auto push to server when script is saved.
 
 ## [1.0.0]